moved some components around for easier testing
